requirements.func.test.txt
# These packages are required when testing the module
git+https://github.com/F5Networks/f5-openstack-test.git@liberty
pytest-cov
Sphinx==1.3.5
